missing type update

+++ src/ray/pubsub/publisher.h
@@ -100,7 +100,7 @@ class SubscriptionIndex {
   /// Publishes the message to relevant subscribers.
   /// Returns true if there are subscribers listening on the entity key of the message,
   /// returns false otherwise.
-  bool Publish(std::shared_ptr<rpc::PubMessage> pub_message, uint64_t msg_size);
+  bool Publish(std::shared_ptr<rpc::PubMessage> pub_message, size_t msg_size);
 
   /// Adds a new subscriber and the key it subscribes to.
   /// When `key_id` is empty, the subscriber subscribes to all keys.
